// work/ai-trading/src/components/home-page/interface.ts
export interface BACK_TEST {
  bot_name: string;
  from: string;
  to: string;
  banner: {
    boxs: {
      [key: string]: {
        id: number;
        title: string;
        value: number;
      };
    };
    chart: { label: string; value: number }[];
    information: {
      initial_balance: number;
      average_trade: number;
      daily_avg_profit: number;
      drawdown: number;
      monthly_avg_profit: number;
      of_trades: number;
      percent_drawdown: number;
      profit_fator: number;
      return_do_ratio: number;
      sharpe_ratio: number;
      winning_percent: number;
    };
  };
  column_chart: {
    monthly_pln: { month: string; name: string; value: number }[];
    trades_of_weekly: { month: string; name: string; value: number }[];
  };
  performance: {
    apr: string;
    aug: string;
    dec: string;
    feb: string;
    jan: string;
    jul: string;
    jun: string;
    key: string;
    mar: string;
    may: string;
    nov: string;
    oct: string;
    sep: string;
    year: string;
    ytd: string;
  }[];
  pie_chart: {
    long_short_ratio: { type: string; value: number }[];
    pln_buy: { type: string; value: number }[];
    pln_sell: { type: string; value: number }[];
  };
  transaction: {
    average_loss: number;
    average_win: number;
    gross_loss: number;
    gross_profit: number;
    largest_loss: number;
    largest_win: number;
    max_consec_loss: number;
    max_consec_win: number;
    of_cancelled_expired: number;
    of_losses: number;
    of_wins: number;
  };
}